Games List: Add checks for skipping invalid games
Some games in the Lutris `pga.db` are missing data,
in most cases this is a result of a game that we
don't want to display on the games list.

Add a method to the Games List dialog that will check
for and skip games missing certain pieces of info,
such as the game runner and install_dir.

When querying the Lutris DB, it appears that only
not-installed Steam games were missing a runner.
As a result they were not being skipped by the
Steam runner skip check, and we were getting
a whole bunch of errors for trying to do operations
on NoneType.

Valid games for the Games List should have a runner
and an install_dir if they are actually installed.
So we skip if they don't meet this criteria, as well
as continuing to skip if a game is a Steam game.

We also add some more None-safe logic to
our installed_at parsing. This also probably
will never be missing but there is no harm in accounting
for it being missing.