Toggle Mark Global button visibility after pressing
If the current tool is successfully marked as global,
hide the Mark Global button using the same logic used
to show/hide the button when the CtInfo dialog is opened.

Also move the show/hide logic for the Mark Global button
to `update_game_list_ui` so that it is not
launcher-specific; `is_mark_global_available`
handles the launcher checks for us.