DXVK Nightly: Simplify by Inheriting from DXVK Ctmod (#509)
* DXVK: Refactor to use __get_data method

Same idea to what we do in GE-Proton ctmod,
allows us to override the data and extract
path in child classes

* DXVK: Refactor extraction to `self.__extract` method

* DXVK: Type hinting

* DXVK Nightly: Simplify by Inheriting DXVK

* DXVK Nightly: Fix fetching Windows MSVC Builds

* DXVK Nightly: Refactor to use commit hash again instead of workflow ID

This allows us to display the commit hash in the releases dropdown
which helps with readibility, and it also allows us to preserve
the functionality of the `get_info_url` method which relies on
using the commit hash.

* DXVK Nightly: Only include workflow runs from `master`

* DXVK Nightly: Refactor extraction

The logic specifically for extracting DXVK Nightly (ZIP)
now lives inside of the DXVK Nightly ctmod.

The main DXVK Ctmod now only knows how to extract itself.