pupgui2: Display Proton versions for Lutris
By default we only display tools in the Lutris `install_directory()` (which
only includes tools inside of `runners/wine`), plus DXVK and vkd3d.

The pattern we use to include DXVK and vkd3d, which are in the
`runtime/dxvk` and `runtime/vkd3d` folders respectively, can also be used
to display Proton tools found in `runners/proton`.

This should also work for other Proton versions so long as they have "Proton"
in their name. If there is a Proton-based compatibility tool enabled for Lutris
that does not include "Proton" in its name, we would need to add another
case to display it using the same pattern we do now, but with its tool name.